# Cadastro_Clientes/target/classes/swagger.yaml
swagger: '2.0'
info:
  description: Microserviço responsável pelo cadastro de contas cartão de clientes
  version: 1.0.0
  title: Microservico Cadastro Clientes
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: localhost:8080
basePath: /cadastro-cliente-api/v1
tags:
  - name: Cliente
    description: Operações relacionadas ao cadastro de clientes
  - name: Solicitacao
    description: Operações de consulta das solicitações de cadastro de contas cartão feitas por clientes
schemes:
  - http
paths:
  /clientes:
    put:
      tags:
        - Cliente
      summary: Cria ou edita os dados de cadastro de um cliente, criando tambem uma nova solicitação
      operationId: salvaCliente
      consumes:
        - "application/json"
      parameters:
        - in: body
          name: clienteRequest
          description: "Dados necessários para realizar o cadastro de um cliente"
          required: true
          schema:
            $ref: '#/definitions/ClienteRequest'
      responses:
        '200':
          description: Cliente cadastrado com sucesso
        '500':
          description: Erro Interno ao Cadastrar Cliente
        '400':
          description: Request Inválida

    get:
      tags:
        - Cliente
      summary: Obtem os dados de cadastro de um Cliente
      operationId: obtemDadosCliente
      produces:
        - "application/json"
      parameters:
        - name: "cpf_cnpj"
          in: "header"
          description: "CPF/CNPJ do Cliente"
          required: true
          type: "string"
      responses:
        '200':
          description: "Sucesso ao pesquisar dados do Cliente"
          schema:
            $ref: "#/definitions/ClienteResponse"
        '500':
          description: Erro Interno ao Cadastrar Cliente
        '400':
          description: Request Inválida
        '404':
          description: Não foi possível encontrar os dados do Cliente

    patch:
      tags:
        - Cliente
      summary: Faz atualização do status do cadastro do Cliente e da conta cartão
      operationId: atualizaStatusCliente
      produces:
        - "application/json"
      parameters:
        - name: "cpf_cnpj"
          in: "header"
          description: "CPF/CNPJ do Cliente"
          required: true
          type: "string"
        - name: "conta_cartao"
          in: "header"
          description: "Conta Cartão do Cliente"
          required: false
          type: "string"
        - name: "status"
          in: "header"
          description: "Status do cadastro do Cliente"
          required: false
          type: string
          enum:
            - CRIADA
            - ANALISE_PENDENTE
            - REJEITADA
      responses:
        '204':
          description: "Status do Cadastro da Conta Cartão do Cliente atualizado com Sucesso"
        '500':
          description: Erro Interno ao Atualizar Status do Cadastro da Conta Cartão do Cliente
        '400':
          description: Request Inválida
        '404':
          description: Nao foi possível encontrar os dados do Cliente

  /solicitacoes:
    patch:
      tags:
        - Solicitacao
      summary: Faz atualização do status da Solicitação de cadastro do Cliente
      operationId: atualizaStatusSolicitacaoCliente
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "header"
          description: "Id da Solicitacao"
          required: true
          type: integer
          format: int64

        - name: "status"
          in: "header"
          description: "Status da Solicitação de cadastro do Cliente"
          required: false
          type: string
          enum:
            - APROVADA
            - REPROVADA
            - EM_ANALISE_DOCUMENTOS
            - EM_ANALISE_FRAUDE
            - EM_ANALISE_CREDITO
      responses:
        '204':
          description: "Status da Solicitação de Cadastro do Cliente atualizado com Sucesso"
        '500':
          description: Erro Interno ao Atualizar Status da Solicitação de Cadastro do Cliente
        '400':
          description: Request Inválida
        '404':
          description: Nao foi possível encontrar os dados do Cliente

    get:
      tags:
        - Solicitacao
      summary: Obtem os dados de solicitações de um Cliente para um periodo específico
      operationId: obtemSolicitacoesClienteNoPeriodo
      produces:
        - "application/json"
      parameters:
        - name: "cpf_cnpj"
          in: "header"
          description: "CPF/CNPJ do Cliente"
          required: true
          type: "string"
        - name: "data_inicial"
          in: "header"
          description: "Data inicial para a busca das solicitações - Padrao dd/MM/YYYY"
          required: true
          type: string
        - name: "data_final"
          in: "header"
          description: "Data final para a busca das solicitações - Padrao dd/MM/YYYY"
          required: true
          type: string
      responses:
        '200':
          description: "Sucesso ao pesquisar dados de Solicitações do Cliente"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/SolicitacaoCadastroResponse"
        '500':
          description: Erro Interno ao Cadastrar Cliente
        '400':
          description: Request Inválida
        '404':
          description: Não foi possível encontrar os dados de Solicitações do Cliente

definitions:
  ClienteRequest:
    type: object
    required:
      - nome
      - cpf_cnpj
      - numero_rg
      - cidade
      - estado
      - pais
      - nome_imagem_rg
      - nome_imagem_cpf_cnpj
    properties:
      nome:
        type: string
        description: Nome Completo do Cliente
        minLength: 1
      cpf_cnpj:
        type: string
        description: CPF/CNPJ do Cliente
        minLength: 11
        maxLength: 14
      numero_rg:
        type: string
        description: Número do RG do Cliente
        minLength: 1
      cidade:
        type: string
        description: Cidade do Cliente
        minLength: 1
      estado:
        type: string
        description: Estado do Cliente
        minLength: 1
      pais:
        type: string
        description: País do Cliente
        minLength: 1
      latitude:
        type: number
        format: double
        description: Latitude do Cliente no momento da solicitação
      longitude:
        type: number
        format: double
        description: Longitude do Cliente no momento da solicitação
      nome_imagem_rg:
        type: string
        description: Nome da imagem do RG do Cliente salva no S3
        minLength: 1
      nome_imagem_cpf_cnpj:
        type: string
        description: Nome da imagem do CPF do Cliente salva no S3
        minLength: 1

  ClienteResponse:
    type: object
    properties:
      nome:
        type: string
        description: Nome Completo do Cliente
      cpf_cnpj:
        type: string
        description: CPF/CNPJ do Cliente
      numero_rg:
        type: string
        description: Número do RG do Cliente
      conta_cartao:
        type: string
        description: Conta Cartão do Cliente
      status_criacao_conta:
        type: string
        description: Status do Cadastro do Cliente
      cidade:
        type: string
        description: Cidade do Cliente
      estado:
        type: string
        description: Estado do Cliente
      pais:
        type: string
        description: País do Cliente

  SolicitacaoCadastroResponse:
    type: object
    properties:
      id:
        type: integer
        format: int64
        description: Id da Solicitaçao
      data_criacao:
        type: string
        description: Data de Criação da Solicitação
      cpf_cnpj_cliente:
        type: string
        description: CPF/CNPJ do Cliente
      status_solicitacao:
        type: string
        description: Status da Solicitação
      latitude:
        type: number
        format: double
        description: Latitude do Cliente no momento da solicitação
      longitude:
        type: number
        format: double
        description: Longitude do Cliente no momento da solicitação
      nome_imagem_rg:
        type: string
        description: Nome da imagem do RG do Cliente salva no S3
      nome_imagem_cpf_cnpj:
        type: string
        description: Nome da imagem do CPF do Cliente salva no S3