# Cadastro_Clientes/target/generated-sources/swagger/src/target/java/main/io/swagger/api/ClientesApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.2.3).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.ClienteRequest;
import io.swagger.model.ClienteResponse;

import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import java.util.List;
import javax.validation.constraints.*;
import javax.validation.Valid;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2020-04-18T17:16:52.388-03:00")

@Api(value = "clientes", description = "the clientes API")
public interface ClientesApi {

    @ApiOperation(value = "Faz atualização do status do cadastro do Cliente e da conta cartão", notes = "", response = Void.class, tags={ "Cliente", })
    @ApiResponses(value = { 
        @ApiResponse(code = 204, message = "Status do Cadastro da Conta Cartão do Cliente atualizado com Sucesso", response = Void.class),
        @ApiResponse(code = 400, message = "Request Inválida", response = Void.class),
        @ApiResponse(code = 404, message = "Nao foi possível encontrar os dados do Cliente", response = Void.class),
        @ApiResponse(code = 500, message = "Erro Interno ao Atualizar Status do Cadastro da Conta Cartão do Cliente", response = Void.class) })
    
    @RequestMapping(value = "/clientes",
        produces = { "application/json" }, 
        method = RequestMethod.PATCH)
    ResponseEntity<Void> atualizaStatusCliente(@ApiParam(value = "CPF/CNPJ do Cliente" ,required=true) @RequestHeader(value="cpf_cnpj", required=true) String cpfCnpj,@ApiParam(value = "Conta Cartão do Cliente" ) @RequestHeader(value="conta_cartao", required=false) String contaCartao,@ApiParam(value = "Status do cadastro do Cliente" , allowableValues="CRIADA, ANALISE_PENDENTE, REJEITADA") @RequestHeader(value="status", required=false) String status);


    @ApiOperation(value = "Obtem os dados de cadastro de um Cliente", notes = "", response = ClienteResponse.class, tags={ "Cliente", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Sucesso ao pesquisar dados do Cliente", response = ClienteResponse.class),
        @ApiResponse(code = 400, message = "Request Inválida", response = Void.class),
        @ApiResponse(code = 404, message = "Não foi possível encontrar os dados do Cliente", response = Void.class),
        @ApiResponse(code = 500, message = "Erro Interno ao Cadastrar Cliente", response = Void.class) })
    
    @RequestMapping(value = "/clientes",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<ClienteResponse> obtemDadosCliente(@ApiParam(value = "CPF/CNPJ do Cliente" ,required=true) @RequestHeader(value="cpf_cnpj", required=true) String cpfCnpj);


    @ApiOperation(value = "Cria ou edita os dados de cadastro de um cliente, criando tambem uma nova solicitação", notes = "", response = Void.class, tags={ "Cliente", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Cliente cadastrado com sucesso", response = Void.class),
        @ApiResponse(code = 400, message = "Request Inválida", response = Void.class),
        @ApiResponse(code = 500, message = "Erro Interno ao Cadastrar Cliente", response = Void.class) })
    
    @RequestMapping(value = "/clientes",
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    ResponseEntity<Void> salvaCliente(@ApiParam(value = "Dados necessários para realizar o cadastro de um cliente" ,required=true )  @Valid @RequestBody ClienteRequest clienteRequest);

}